Refactor delegator activities
