+++ rust/processor/src/db/common/models/stake_models/mod.rs
@@ -1,3 +1,4 @@
+pub mod delegator_activities;
 pub mod delegator_balances;
 pub mod delegator_pools;
 pub mod stake_utils;
