+++ rust/processor/src/db/postgres/models/stake_models/delegator_activities.rs
@@ -3,15 +3,12 @@
 // This is required because a diesel macro makes clippy sad
 #![allow(clippy::extra_unused_lifetimes)]
 
-use crate::db::common::models::stake_models::stake_utils::StakeEvent;
 use crate::{
-    schema::delegated_staking_activities,
-    utils::{
-        counters::PROCESSOR_UNKNOWN_TYPE_COUNT,
-        util::{standardize_address, u64_to_bigdecimal},
+    db::common::models::stake_models::delegator_activities::{
+        RawDelegatedStakingActivity, RawDelegatedStakingActivityConvertible,
     },
+    schema::delegated_staking_activities,
 };
-use aptos_protos::transaction::v1::{transaction::TxnData, Transaction};
 use bigdecimal::BigDecimal;
 use field_count::FieldCount;
 use serde::{Deserialize, Serialize};
@@ -28,82 +25,15 @@ pub struct DelegatedStakingActivity {
     pub amount: BigDecimal,
 }
 
-impl DelegatedStakingActivity {
-    /// Pretty straightforward parsing from known delegated staking events
-    pub fn from_transaction(transaction: &Transaction) -> anyhow::Result<Vec<Self>> {
-        let mut delegator_activities = vec![];
-        let txn_data = match transaction.txn_data.as_ref() {
-            Some(data) => data,
-            None => {
-                PROCESSOR_UNKNOWN_TYPE_COUNT
-                    .with_label_values(&["DelegatedStakingActivity"])
-                    .inc();
-                tracing::warn!(
-                    transaction_version = transaction.version,
-                    "Transaction data doesn't exist",
-                );
-                return Ok(delegator_activities);
-            },
-        };
-
-        let txn_version = transaction.version as i64;
-        let events = match txn_data {
-            TxnData::User(txn) => &txn.events,
-            TxnData::BlockMetadata(txn) => &txn.events,
-            TxnData::Validator(txn) => &txn.events,
-            _ => return Ok(delegator_activities),
-        };
-        for (index, event) in events.iter().enumerate() {
-            let event_index = index as i64;
-            if let Some(staking_event) =
-                StakeEvent::from_event(event.type_str.as_str(), &event.data, txn_version)?
-            {
-                let activity = match staking_event {
-                    StakeEvent::AddStakeEvent(inner) => DelegatedStakingActivity {
-                        transaction_version: txn_version,
-                        event_index,
-                        delegator_address: standardize_address(&inner.delegator_address),
-                        pool_address: standardize_address(&inner.pool_address),
-                        event_type: event.type_str.clone(),
-                        amount: u64_to_bigdecimal(inner.amount_added),
-                    },
-                    StakeEvent::UnlockStakeEvent(inner) => DelegatedStakingActivity {
-                        transaction_version: txn_version,
-                        event_index,
-                        delegator_address: standardize_address(&inner.delegator_address),
-                        pool_address: standardize_address(&inner.pool_address),
-                        event_type: event.type_str.clone(),
-                        amount: u64_to_bigdecimal(inner.amount_unlocked),
-                    },
-                    StakeEvent::WithdrawStakeEvent(inner) => DelegatedStakingActivity {
-                        transaction_version: txn_version,
-                        event_index,
-                        delegator_address: standardize_address(&inner.delegator_address),
-                        pool_address: standardize_address(&inner.pool_address),
-                        event_type: event.type_str.clone(),
-                        amount: u64_to_bigdecimal(inner.amount_withdrawn),
-                    },
-                    StakeEvent::ReactivateStakeEvent(inner) => DelegatedStakingActivity {
-                        transaction_version: txn_version,
-                        event_index,
-                        delegator_address: standardize_address(&inner.delegator_address),
-                        pool_address: standardize_address(&inner.pool_address),
-                        event_type: event.type_str.clone(),
-                        amount: u64_to_bigdecimal(inner.amount_reactivated),
-                    },
-                    StakeEvent::DistributeRewardsEvent(inner) => DelegatedStakingActivity {
-                        transaction_version: txn_version,
-                        event_index,
-                        delegator_address: "".to_string(),
-                        pool_address: standardize_address(&inner.pool_address),
-                        event_type: event.type_str.clone(),
-                        amount: u64_to_bigdecimal(inner.rewards_amount),
-                    },
-                    _ => continue,
-                };
-                delegator_activities.push(activity);
-            }
+impl RawDelegatedStakingActivityConvertible for DelegatedStakingActivity {
+    fn from_raw(raw: RawDelegatedStakingActivity) -> Self {
+        Self {
+            transaction_version: raw.transaction_version,
+            event_index: raw.event_index,
+            delegator_address: raw.delegator_address,
+            pool_address: raw.pool_address,
+            event_type: raw.event_type,
+            amount: raw.amount,
         }
-        Ok(delegator_activities)
     }
 }
